add logic behind Detail & Delete button

+++ Mvc_Project Db/Controllers/EmployeeController.cs
@@ -57,6 +57,37 @@ namespace Mvc_Project_Db.Controllers
             }
             return View(employeeIndb);
         }
+        [HttpPost]
+        public ActionResult Edit(Employee employee)
+        {
+            if (employee == null)
+            {
+                return HttpNotFound();
+            }
+            if (!ModelState.IsValid)
+            {
+                return View(employee);
+            }
+            var employeeIndb = context.Employees.Find(employee.Id);
+            if (employeeIndb == null)
+            {
+                return HttpNotFound();
+            }
+            employeeIndb.Name = employee.Name;
+            employeeIndb.Address = employee.Address;
+            employeeIndb.salary = employee.salary;
+            context.SaveChanges();
+            return RedirectToAction("Index");
+        }
+        public ActionResult Details(int id)
+        {
+            var employeeIndb = context.Employees.Find(id);
+            if (employeeIndb == null)
+            {
+                return HttpNotFound();
+            }
+            return View(employeeIndb);
+        }
         public ActionResult Delete(int id)
         {
             var employeeIndb = context.Employees.Find(id);
@@ -64,14 +95,28 @@ namespace Mvc_Project_Db.Controllers
             {
                 return HttpNotFound();
             }
+            return View(employeeIndb);
+        }
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Delete(Employee employee)
+        {
+            if (employee == null)
+            {
+                return HttpNotFound();
+            }
+
+            var employeeIndb = context.Employees.Find(employee.Id);
+            if (employeeIndb == null)
+            {
+                return HttpNotFound();
+            }
+
             context.Employees.Remove(employeeIndb);
             context.SaveChanges();
+
             return RedirectToAction("Index");
         }
-        public ActionResult EditAll()
-        {
-            var employeelist = context.Employees.ToList();
-            return View(employeelist);
-        }
+
     }
 }
